Namespace the ansible.utils.display logging configuration
The logging configuration provided by ansible.utils.display was not
namespaced which could cause conflicts with other python software
running with Ansible such as plugins, modules, etc.

This patch is effectively no-op in the sense that it keeps the exact
same formatting and configuration but will prevent conflicts from
happening.
It also makes it easier to customize the logging configuration in
the future if we need to or would like to expose a way for users
to provide their own configuration.

Fixes: #41838
Fixes: #25757
Fixes: #25759